requirements.txt
python-dotenv
keyboard
requests
google-generativeai
opencv-python
Pillow
pygame